requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langgraph>=0.0.15
tavily-python>=0.2.6
python-dotenv>=1.0.0
langchain-core>=0.1.0
typing-extensions>=4.8.0